// packages/sdk/src/lib/abi/Inbox.ts
export type InboxMessageDeliveredEvent = TypedEvent<
  [BigNumber, string],
  { messageNum: BigNumber; data: string }
>;