// packages/sdk/src/lib/abi/Bridge.ts
export type MessageDeliveredEvent = TypedEvent<
  [BigNumber, string, string, number, string, string, BigNumber, BigNumber],
  {
    messageIndex: BigNumber;
    beforeInboxAcc: string;
    inbox: string;
    kind: number;
    sender: string;
    messageDataHash: string;
    baseFeeL1: BigNumber;
    timestamp: BigNumber;
  }
>;